+++ PyCurrentWindow.py
@@ -1,0 +1,4 @@
+from pycurrentwindow import main
+
+if __name__ == "__main__":
+    main()
