+++ pycurrentwindow/utils/app_data.py
@@ -2,7 +2,8 @@ from os import getenv, makedirs, path
 
 
 def app_data() -> str:
-    """Returns location of the AppData directory for the app , creating it if it does not exist
+    """Returns location of the AppData directory for the app,
+    creating it if it does not exist
 
     Returns:
         str: The full file location
